test(provider): move capabilities integration test and fix import cycle
- migrate the ModelProviderGateway integration test to
  ProviderCapabilitiesRepository under tests/integration/http
- import ProviderClient / adapter classes from their submodules to break
  the circular import introduced by moving the repository into the
  provider package
